Move Loading to separate component
- Use User typings from utils
