+++ src/Root/App/Multiplexer/index.tsx
@@ -5,51 +5,13 @@ import { _cs } from '@togglecorp/fujs';
 
 import Navbar from '#components/Navbar';
 import DomainContext from '#components/DomainContext';
+import Loading from '#components/Loading';
+
+import { User } from '#utils/typings';
 
 import routeSettings from './route';
 
 import styles from './styles.css';
-
-interface LoadingProps {
-    message: string;
-    delay?: number;
-}
-function Loading(props: LoadingProps) {
-    const {
-        message,
-        delay = 200,
-    } = props;
-
-    const initialVisibility = delay <= 0;
-    const [visibility, setVisibility] = useState(initialVisibility);
-
-    useEffect(
-        () => {
-            const timeout = setTimeout(
-                () => {
-                    setVisibility(true);
-                },
-                delay,
-            );
-            return () => {
-                clearTimeout(timeout);
-            };
-        },
-        [delay],
-    );
-
-    return (
-        <div className={styles.loading}>
-            {visibility ? <h3>{message}</h3> : undefined}
-        </div>
-    );
-}
-
-interface User {
-    username: string;
-    email: string;
-    id: number;
-}
 
 interface Me {
     me?: User;
