+++ src/Root/App/Multiplexer/styles.css
@@ -1,10 +1,3 @@
-.loading {
-    display: flex;
-    align-items: center;
-    flex-grow: 1;
-    justify-content: center;
-}
-
 .multiplexer {
     display: flex;
     position: relative;
