+++ src/components/Loading/styles.css
@@ -1,0 +1,7 @@
+.loading {
+    display: flex;
+    align-items: center;
+    flex-grow: 1;
+    justify-content: center;
+}
+
